Sets up CI workflow
Creates a CI workflow to automate infrastructure deployment and server configuration.

This workflow uses Terraform to provision infrastructure on AWS and Ansible to configure the server with Nginx.
It includes steps for:
- Checking out the code
- Configuring AWS credentials
- Installing Terraform and Ansible
- Creating SSH keys and importing them to AWS
- Initializing, planning, and applying Terraform
- Retrieving Terraform outputs
- Configuring SSH for Ansible
- Creating an Ansible inventory
- Executing the Ansible playbook

The Terraform workflow outputs the public IP of the created instance, which is then used by Ansible to configure the server.